ref: refactor items in some components
