+++ frontend/src/components/Footer/items.js
@@ -1,0 +1,8 @@
+const items = [
+  { link: "https://www.facebook.com/projetoales", icon: "fab fa-facebook-f" },
+  { link: "https://www.instagram.com/projetoales/", icon: "fab fa-instagram" },
+  { link: "https://www.youtube.com/channel/UCXL_iYkkNcNpWDE6inM_LTg", icon: "fab fa-youtube" },
+  { link: "https://github.com/Projeto-ALES", icon: "fab fa-github-alt" },
+];
+
+export default items;
