+++ frontend/src/components/SideBar/items.js
@@ -1,0 +1,8 @@
+const items = [
+  { id: 1, text: "Home", icon: "fas fa-home" },
+  { id: 2, text: "Localização", icon: "fas fa-map-marked-alt" },
+  { id: 3, text: "Matérias", icon: "far fa-calendar-check" },
+  { id: 4, text: "Contato", icon: "fas fa-paper-plane" },
+];
+
+export default items;
