+++ frontend/src/views/Home/Sections/WhatDo/activities.js
@@ -1,0 +1,7 @@
+const activities = [
+  { id: 1, text: "UPA", icon: "fas fa-graduation-cap" },
+  { id: 2, text: "MUPA", icon: "fas fa-landmark" },
+  { id: 3, text: "Hackathons", icon: "fas fa-users" },
+];
+
+export default activities;
