+++ frontend/src/views/Home/Sections/WhatDo/courses.js
@@ -1,0 +1,7 @@
+const courses = [
+  { id: 1, text: "Programação", icon: "fas fa-keyboard" },
+  { id: 2, text: "Robótica", icon: "fas fa-robot" },
+  { id: 3, text: "Desenho", icon: "fas fa-palette" },
+];
+
+export default courses;
